requirements.txt
fastapi 
uvicorn
requests
python-dotenv
faster-whisper
yt-dlp
ffmpeg-python
aiohttp
logging